+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 # oneLiner Changelog
 
-## [Unreleased]
+## [0.0.2]
+### Added
+- show lines from default file path and register pageUp and pageDown actions.
+- first useful version
+
 
 ## [0.0.1]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
-- show default string at status bar
+- show default string at status bar.